fix: proto-implementations
The current proto-implementations are wrong, because we don't want to
obscure methods that need be implemented by _implementing_ them, also
when the implementation is just an error message.

This commit removes the proto-implementations that need actually be
implemented by instances of Menge, while adding documentation for the
proto-implementations that are somewhat complex.